requirements.txt
sentence-transformers 
faiss-cpu 
langchain 
unstructured 
langchain_mistralai 
langchain_huggingface
huggingface_hub
streamlit
langchain-community
langchain-google-genai
langgraph
python-dotenv